* refactor tileentitylist out of map_t, map_t's cause stack overflows during loadMap()
